# frontend/openapi/govos-api.yaml
openapi: 3.0.3
info:
  title: GovOS API
  version: 0.1.0
  description: GovOS Platform REST API
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/auth/refresh:
    post:
      tags: [Authentication]
      operationId: refresh
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseLoginResponse'
...
components:
  schemas:
...
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
...
    AuthUserResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
        roles:
          type: array
          items:
            type: string
        permissions:
          type: array
          items:
            type: string
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          format: int64
        tokenType:
          type: string
        user:
          $ref: '#/components/schemas/AuthUserResponse'
...
    ApiResponseLoginResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/LoginResponse'
        message:
          type: string
        timestamp:
          type: string
        requestId:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []